Move event pruner to a separate module.
